==> modulos/calculos/00_pool_bases.ipynb <==
import pandas as pd
import warnings
import os
warnings.filterwarnings('ignore')

script_dir = os.getcwd()
base_path = os.path.abspath(os.path.join(script_dir, '..', '..'))

%run {os.path.join(base_path, 'modulos/funciones/00_funciones_pool_bases.ipynb')}

relative_path = 'data_input/EPH/personas_total_urbano/'
full_path = os.path.join(base_path, relative_path)

filename = 'personas_tot_urb_3T_21.txt'
filename = 'personas_tot_urb_3T_22.txt'

filename = 'hogar_tot_urb_3T_2021.txt'
filename = 'hogar_tot_urb_3T_2022.txt'

# Read data

df_people_2021 = pd.read_csv(full_path + filename, delimiter=';', low_memory=False)
df_people_2022 = pd.read_csv(full_path + filename, delimiter=';', low_memory=False)

df_houses_2021 = pd.read_csv(full_path + filename, delimiter=';', low_memory=False)
df_houses_2022 = pd.read_csv(full_path + filename, delimiter=';', low_memory=False)

conversion_factor = 0.5458771634
df_merged_without_copies_people = merge_and_calculate_people(df_people_2021, df_people_2022, 2021, 2022, conversion_factor)
df_merged_without_copies_people.to_csv(pathdata + 'data_output/Base_pool_individuos_solo_con_replicas_actuales.csv')

df_merged_without_copies_houses = merge_and_calculate_houses(df_houses_2021, df_houses_2022, 2021, 2022, conversion_factor)
df_merged_without_copies_houses.to_csv(pathdata + 'data_output/Base_pool_hogares_solo_con_replicas_actuales.csv')
