==> modulos/calculos/04_computo_CBT.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
warnings.filterwarnings('ignore')

pathdata = '/home/daniu/Documentos/fundar/indice-mercado-trabajo-ingresos/'
pathdata = '/Users/danielarisaro/Documents/fundar/indice-mercado-trabajo-ingresos/'
pathdata = '/home/daniufundar/Documents/Fundar/indice-mercado-trabajo-ingresos/'

filename = 'data_input/canastas_basicas/canasta-basica-total-regiones-del-pais.csv'

df = pd.read_csv(pathdata + filename)

df.index = pd.to_datetime(df['indice_tiempo'])
df = df.drop(['indice_tiempo'], axis=1)
df_rolling = df.rolling(window=3, center=True).mean()
df_CBT = df_rolling.loc[df_rolling.index.month.isin([1,4,7,10])]

# Define the mapping dictionary
mapping = {
    1: lambda x: f'1T{x.year}',
    4: lambda x: f'2T{x.year}',
    7: lambda x: f'3T{x.year}',
    10: lambda x: f'4T{x.year}'
}

# Add a new column with the renamed values
df_CBT.loc[:, 'Trimestre'] = df_CBT.index.map(lambda x: mapping.get(x.month, lambda x: x)(x))
df_CBT = df_CBT.dropna()

df_CBT.to_csv('/home/daniufundar/Documents/Fundar/indice-mercado-trabajo-ingresos/data_output/Canasta_Basica_Total_Regiones_2016-2022-promedios-moviles.csv')
