==> modulos/calculos/06_tabla_caracteristicas_EPH.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from matplotlib.patches import Rectangle
import warnings
warnings.filterwarnings('ignore')

# Read data
pathdata = '/home/daniu/Documentos/fundar/indice-mercado-trabajo-ingresos/'
pathdata = '/home/daniufundar/Documents/Fundar/indice-mercado-trabajo-ingresos/'
pathdata = '/Users/danielarisaro/Documents/fundar/indice-mercado-trabajo-ingresos/'

df_people_pool = pd.read_csv(pathdata + 'data_output/Base_pool_individuos_solo_con_replicas_actuales.csv', low_memory=False, index_col=0)
df_houses_pool = pd.read_csv(pathdata + 'data_output/Base_pool_hogares_solo_con_replicas_actuales.csv', low_memory=False, index_col=0)

df_people_2021 = pd.read_csv(pathdata + 'data_input/personas_tot_urb_3T_21.txt', delimiter=';', low_memory=False)
df_people_2022 = pd.read_csv(pathdata + 'data_input/personas_tot_urb_3T_22.txt', delimiter=';', low_memory=False)

df_houses_2021 = pd.read_csv(pathdata + 'data_input/hogar_tot_urb_3T_2021.txt', delimiter=';', low_memory=False)
df_houses_2022 = pd.read_csv(pathdata + 'data_input/hogar_tot_urb_3T_2022.txt', delimiter=';', low_memory=False)

def row_counts(df_year1, df_year2, year1, year2):
    count_year1 = len(df_year1)
    count_year2 = len(df_year2)
    merged_df = pd.merge(df_year1, df_year2, on=['CODUSU', 'COMPONENTE'], how='inner')
    count_both_years = len(merged_df)

    # Calculate total population
    pob_year1 = df_year1['PONDERA'].sum()
    pob_year2 = df_year2['PONDERA'].sum()

    # Calculate number of surveys performed
    encuesta_realizada_year1 = df_year1[df_year1['H15'] != 2]['H15'].count()
    pob_encuesta_realizada_year1 = df_year1[df_year1['H15'] != 2]['PONDERA'].sum()
    encuesta_realizada_year2 = df_year2[df_year2['H15'] != 2]['H15'].count()
    pob_encuesta_realizada_year2 = df_year2[df_year2['H15'] != 2]['PONDERA'].sum()

    # Calculate number of imputed surveys
    encuesta_imputada_year1 = df_year1[df_year1['IMPUTA'] == 2]['IMPUTA'].count()
    pob_encuesta_imputada_year1 = df_year1[df_year1['IMPUTA'] == 2]['PONDERA'].sum()
    encuesta_imputada_year2 = df_year2[df_year2['IMPUTA'] == 2]['IMPUTA'].count()
    pob_encuesta_imputada_year2 = df_year2[df_year2['IMPUTA'] == 2]['PONDERA'].sum()

    # Create the table as a pandas dataframe
    data = {
        'Anio': [year1, year2],
        'Cantidad de observaciones': [count_year1, count_year2],
        'Observaciones en ambos anios': [count_both_years, count_both_years],
        'Poblacion total': [pob_year1, pob_year2],
        'Encuestas realizadas en el anio': [encuesta_realizada_year1, encuesta_realizada_year2],
        'Poblacion de encuestas realizadas en el anio': [pob_encuesta_realizada_year1, pob_encuesta_realizada_year2],
        'Encuestas imputadas en el anio': [encuesta_imputada_year1, encuesta_imputada_year2],
        'Poblacion de encuestas imputadas en el anio': [pob_encuesta_imputada_year1, pob_encuesta_imputada_year2]
    }
    table = pd.DataFrame(data)
    table.set_index('Anio', inplace=True)

    return table

result_table = row_counts(df_people_2021, df_people_2022, 2021, 2022)
result_table


## Calculo inactivos, desocupados y ocupados

base_pool_inactivos = df_people_pool[(df_people_pool['CH06'] > 14) & (df_people_pool['ESTADO'] == 3)]['PONDERA_new'].sum().round(0)
base_pool_desocupados = df_people_pool[(df_people_pool['CH06'] > 14) & (df_people_pool['ESTADO'] == 2)]['PONDERA_new'].sum().round(0)
base_pool_ocupados = df_people_pool[(df_people_pool['CH06'] > 14) & (df_people_pool['ESTADO'] == 1)]['PONDERA_new'].sum().round(0)

base_2022_inactivos = df_people_2022[(df_people_2022['CH06'] > 14) & (df_people_2022['ESTADO'] == 3)]['PONDERA'].sum().round(0)
base_2022_desocupados = df_people_2022[(df_people_2022['CH06'] > 14) & (df_people_2022['ESTADO'] == 2)]['PONDERA'].sum().round(0)
base_2022_ocupados = df_people_2022[(df_people_2022['CH06'] > 14) & (df_people_2022['ESTADO'] == 1)]['PONDERA'].sum().round(0)

base_2021_inactivos = df_people_2021[(df_people_2021['CH06'] > 14) & (df_people_2021['ESTADO'] == 3)]['PONDERA'].sum().round(0)
base_2021_desocupados = df_people_2021[(df_people_2021['CH06'] > 14) & (df_people_2021['ESTADO'] == 2)]['PONDERA'].sum().round(0)
base_2021_ocupados = df_people_2021[(df_people_2021['CH06'] > 14) & (df_people_2021['ESTADO'] == 1)]['PONDERA'].sum().round(0)

# Crear un DataFrame
data = {
    'Base pool': [base_pool_inactivos, base_pool_desocupados, base_pool_ocupados],
    'Base 2022': [base_2022_inactivos, base_2022_desocupados, base_2022_ocupados],
    'Base 2021': [base_2021_inactivos, base_2021_desocupados, base_2021_ocupados]
}

df = pd.DataFrame(data, index=['Inactivos', 'Desocupados', 'Ocupados'])
print(df)
